+++ DT_MARS-EXPLORE/Polar_Mission.h
@@ -6,6 +6,6 @@ class Polar_Mission :
 private:
     static int NumOfPMissions;
 public:
-    Polar_Mission(int TLOC_x, int MDUR_x, int SIG_x, int FD_x, int ID_x);
+    Polar_Mission(double TLOC_x, int MDUR_x, int SIG_x, int FD_x, int ID_x);
 };
